List Available Datasets from UCI ML Repository

Prints a list of datasets that can be imported via the fetch_ucirepo function.

Usage

list_available_datasets(filter, search, area)

Arguments

filter

Character. Optional query to filter available datasets based on a label.

search

Character. Optional query to search for available datasets by name.

area

Character. Optional query to filter available datasets based on subject area.

Value

Prints the list of available datasets.

Invisibly returns a data frame containing the list of available datasets with columns of:

id: Integer ID for the data set.

name: Name of Dataset

url: Download location of the data set

Examples

list_available_datasets(search = "iris")
#> -------------------------------------------------------------
#> The following datasets are available for search query "iris":
#> -------------------------------------------------------------
#> Dataset Name    ID    
#> ------------    --    
#> Iris            53    
#> 
list_available_datasets(area = "social science")
#> -------------------------------------
#> The following datasets are available:
#> For subject area: social science
#> -------------------------------------
#> Dataset Name                                        ID    
#> ------------                                        --    
#> Adult                                               2     
#> Balance Scale                                       12    
#> Balloons                                            13    
#> Census Income                                       20    
#> Hayes-Roth                                          44    
#> Nursery                                             76    
#> Congressional Voting Records                        105   
#> US Census Data (1990)                               116   
#> Census-Income (KDD)                                 117   
#> Statlog (German Credit Data)                        144   
#> Communities and Crime                               183   
#> Communities and Crime Unnormalized                  211   
#> Bike Sharing                                        275   
#> Student Performance                                 320   
#> Drug Consumption (Quantified)                       373   
#> Autism Screening Adult                              426   
#> Gender by Name                                      591   
#> Predict Students' Dropout and Academic Success      697   
#> Power Consumption of Tetouan City                   849   
#> Higher Education Students Performance Evaluation    856   
#> 
list_available_datasets(filter = "python") # Required for now...
#> --------------------------------------------
#> The following python datasets are available:
#> --------------------------------------------
#> Dataset Name                                                                            ID    
#> ------------                                                                            --    
#> Abalone                                                                                 1     
#> Adult                                                                                   2     
#> Annealing                                                                               3     
#> Audiology (Standardized)                                                                8     
#> Auto MPG                                                                                9     
#> Automobile                                                                              10    
#> Balance Scale                                                                           12    
#> Balloons                                                                                13    
#> Breast Cancer                                                                           14    
#> Breast Cancer Wisconsin (Original)                                                      15    
#> Breast Cancer Wisconsin (Prognostic)                                                    16    
#> Breast Cancer Wisconsin (Diagnostic)                                                    17    
#> Pittsburgh Bridges                                                                      18    
#> Car Evaluation                                                                          19    
#> Census Income                                                                           20    
#> Chess (King-Rook vs. King-Pawn)                                                         22    
#> Chess (King-Rook vs. King)                                                              23    
#> Connect-4                                                                               26    
#> Credit Approval                                                                         27    
#> Japanese Credit Screening                                                               28    
#> Computer Hardware                                                                       29    
#> Contraceptive Method Choice                                                             30    
#> Covertype                                                                               31    
#> Cylinder Bands                                                                          32    
#> Dermatology                                                                             33    
#> Echocardiogram                                                                          38    
#> Ecoli                                                                                   39    
#> Flags                                                                                   40    
#> Glass Identification                                                                    42    
#> Haberman's Survival                                                                     43    
#> Hayes-Roth                                                                              44    
#> Heart Disease                                                                           45    
#> Hepatitis                                                                               46    
#> Horse Colic                                                                             47    
#> Image Segmentation                                                                      50    
#> Ionosphere                                                                              52    
#> Iris                                                                                    53    
#> ISOLET                                                                                  54    
#> Lenses                                                                                  58    
#> Letter Recognition                                                                      59    
#> Liver Disorders                                                                         60    
#> Lung Cancer                                                                             62    
#> Lymphography                                                                            63    
#> Molecular Biology (Splice-junction Gene Sequences)                                      69    
#> MONK's Problems                                                                         70    
#> Mushroom                                                                                73    
#> Musk (Version 1)                                                                        74    
#> Musk (Version 2)                                                                        75    
#> Nursery                                                                                 76    
#> Page Blocks Classification                                                              78    
#> Optical Recognition of Handwritten Digits                                               80    
#> Pen-Based Recognition of Handwritten Digits                                             81    
#> Post-Operative Patient                                                                  82    
#> Primary Tumor                                                                           83    
#> Servo                                                                                   87    
#> Shuttle Landing Control                                                                 88    
#> Solar Flare                                                                             89    
#> Soybean (Large)                                                                         90    
#> Soybean (Small)                                                                         91    
#> Challenger USA Space Shuttle O-Ring                                                     92    
#> Spambase                                                                                94    
#> SPECT Heart                                                                             95    
#> SPECTF Heart                                                                            96    
#> Tic-Tac-Toe Endgame                                                                     101   
#> Congressional Voting Records                                                            105   
#> Waveform Database Generator (Version 1)                                                 107   
#> Wine                                                                                    109   
#> Yeast                                                                                   110   
#> Zoo                                                                                     111   
#> US Census Data (1990)                                                                   116   
#> Census-Income (KDD)                                                                     117   
#> El Nino                                                                                 122   
#> Statlog (Australian Credit Approval)                                                    143   
#> Statlog (German Credit Data)                                                            144   
#> Statlog (Heart)                                                                         145   
#> Statlog (Landsat Satellite)                                                             146   
#> Statlog (Image Segmentation)                                                            147   
#> Statlog (Shuttle)                                                                       148   
#> Statlog (Vehicle Silhouettes)                                                           149   
#> Connectionist Bench (Sonar, Mines vs. Rocks)                                            151   
#> Cloud                                                                                   155   
#> Poker Hand                                                                              158   
#> MAGIC Gamma Telescope                                                                   159   
#> Mammographic Mass                                                                       161   
#> Forest Fires                                                                            162   
#> Concrete Compressive Strength                                                           165   
#> Ozone Level Detection                                                                   172   
#> Parkinsons                                                                              174   
#> Blood Transfusion Service Center                                                        176   
#> Communities and Crime                                                                   183   
#> Acute Inflammations                                                                     184   
#> Wine Quality                                                                            186   
#> Parkinsons Telemonitoring                                                               189   
#> Cardiotocography                                                                        193   
#> Steel Plates Faults                                                                     198   
#> Communities and Crime Unnormalized                                                      211   
#> Vertebral Column                                                                        212   
#> Bank Marketing                                                                          222   
#> ILPD (Indian Liver Patient Dataset)                                                     225   
#> Skin Segmentation                                                                       229   
#> Individual Household Electric Power Consumption                                         235   
#> Energy Efficiency                                                                       242   
#> Fertility                                                                               244   
#> ISTANBUL STOCK EXCHANGE                                                                 247   
#> User Knowledge Modeling                                                                 257   
#> EEG Eye State                                                                           264   
#> Banknote Authentication                                                                 267   
#> Gas Sensor Array Drift at Different Concentrations                                      270   
#> Bike Sharing                                                                            275   
#> Thoracic Surgery Data                                                                   277   
#> Airfoil Self-Noise                                                                      291   
#> Wholesale customers                                                                     292   
#> Combined Cycle Power Plant                                                              294   
#> Diabetes 130-US Hospitals for Years 1999-2008                                           296   
#> Tennis Major Tournament Match Statistics                                                300   
#> Dow Jones Index                                                                         312   
#> Student Performance                                                                     320   
#> Phishing Websites                                                                       327   
#> Diabetic Retinopathy Debrecen                                                           329   
#> Online News Popularity                                                                  332   
#> Chronic Kidney Disease                                                                  336   
#> Mice Protein Expression                                                                 342   
#> Default of Credit Card Clients                                                          350   
#> Online Retail                                                                           352   
#> Occupancy Detection                                                                     357   
#> Air Quality                                                                             360   
#> Polish Companies Bankruptcy                                                             365   
#> Dota2 Games Results                                                                     367   
#> Facebook Metrics                                                                        368   
#> HTRU2                                                                                   372   
#> Drug Consumption (Quantified)                                                           373   
#> Appliances Energy Prediction                                                            374   
#> Website Phishing                                                                        379   
#> YouTube Spam Collection                                                                 380   
#> Beijing PM2.5                                                                           381   
#> Cervical Cancer (Risk Factors)                                                          383   
#> Stock Portfolio Performance                                                             390   
#> Sales Transactions Weekly                                                               396   
#> Daily Demand Forecasting Orders                                                         409   
#> Autistic Spectrum Disorder Screening Data for Children                                  419   
#> Autism Screening Adult                                                                  426   
#> Absenteeism at work                                                                     445   
#> Breast Cancer Coimbra                                                                   451   
#> Drug Reviews (Druglib.com)                                                              461   
#> Drug Reviews (Drugs.com)                                                                462   
#> Superconductivty Data                                                                   464   
#> Student Academics Performance                                                           467   
#> Online Shoppers Purchasing Intention Dataset                                            468   
#> Electrical Grid Stability Simulated Data                                                471   
#> Real Estate Valuation                                                                   477   
#> Travel Reviews                                                                          484   
#> Travel Review Ratings                                                                   485   
#> Facebook Live Sellers in Thailand                                                       488   
#> Metro Interstate Traffic Volume                                                         492   
#> Hepatitis C Virus (HCV) for Egyptian patients                                           503   
#> Heart Failure Clinical Records                                                          519   
#> Early Stage Diabetes Risk Prediction                                                    529   
#> Pedestrians in Traffic                                                                  536   
#> Cervical Cancer Behavior Risk                                                           537   
#> Estimation of Obesity Levels Based On Eating Habits and Physical Condition              544   
#> Rice (Cammeo and Osmancik)                                                              545   
#> Algerian Forest Fires                                                                   547   
#> Gas Turbine CO and NOx Emission Data Set                                                551   
#> Apartment for Rent Classified                                                           555   
#> Seoul Bike Sharing Demand                                                               560   
#> Iranian Churn                                                                           563   
#> Bone marrow transplant: children                                                        565   
#> COVID-19 Surveillance                                                                   567   
#> HCV data                                                                                571   
#> Taiwanese Bankruptcy Prediction                                                         572   
#> Myocardial infarction complications                                                     579   
#> Student Performance on an Entrance Examination                                          582   
#> Gender by Name                                                                          591   
#> Productivity Prediction of Garment Employees                                            597   
#> AI4I 2020 Predictive Maintenance Dataset                                                601   
#> Dry Bean                                                                                602   
#> In-Vehicle Coupon Recommendation                                                        603   
#> Predict Students' Dropout and Academic Success                                          697   
#> Auction Verification                                                                    713   
#> NATICUSdroid (Android Permissions)                                                      722   
#> Toxicity                                                                                728   
#> DARWIN                                                                                  732   
#> Accelerometer Gyro Mobile Phone                                                         755   
#> Glioma Grading Clinical and Mutation Features                                           759   
#> Multivariate Gait Data                                                                  760   
#> Land Mines                                                                              763   
#> Single Elder Home Monitoring: Gas and Position                                          799   
#> Sepsis Survival Minimal Clinical Records                                                827   
#> Secondary Mushroom                                                                      848   
#> Power Consumption of Tetouan City                                                       849   
#> Raisin                                                                                  850   
#> Steel Industry Energy Consumption                                                       851   
#> Higher Education Students Performance Evaluation                                        856   
#> Risk Factor Prediction of Chronic Kidney Disease                                        857   
#> Maternal Health Risk                                                                    863   
#> Room Occupancy Estimation                                                               864   
#> Cirrhosis Patient Survival Prediction                                                   878   
#> SUPPORT2                                                                                880   
#> National Health and Nutrition Health Survey 2013-2014 (NHANES) Age Prediction Subset    887   
#> AIDS Clinical Trials Group Study 175                                                    890   
#> CDC Diabetes Health Indicators                                                          891   
#> Recipe Reviews and User Feedback                                                        911   
#> Forty Soybean Cultivars from Subsequent Harvests                                        913   
#> Differentiated Thyroid Cancer Recurrence                                                915   
#> Infrared Thermography Temperature                                                       925   
#> National Poll on Healthy Aging (NPHA)                                                   936   
#> Regensburg Pediatric Appendicitis                                                       938   
#> RT-IoT2022                                                                              942   
#> PhiUSIIL Phishing URL (Website)                                                         967   
#>